- Gerando gráfico para o artigo
