+++ code/main.py
@@ -2,9 +2,11 @@ import math
 import os, re
 import threading
 import pandas as pd
+import matplotlib.pyplot as plt
+
 from util.preprocess import PrepData, AutoStem
 from language_clf import LangClf
-from util.util import count_words, stem_text
+from util.util import count_words, stem_text, draw_plot
 import pprint
 from nltk import RegexpTokenizer
 
@@ -15,23 +17,6 @@ stems_path = r'../Resources/stems/'
 data_list = os.listdir(path)
 prep = PrepData(path)
 dataset = prep.get_datasets()
-
-
-def stem_words(name):
-    data = pd.read_csv(path + name)
-    label = name.replace('.csv', '')
-    print('Processing: ', label)
-    scripture = data['Scripture']
-
-    stem = AutoStem(scripture)
-    stem.freq_counter()
-
-    stem.stem_words()
-    data = {label: list(filter(lambda x: type(x) == str, stem.select_stem()))}
-
-    df = pd.DataFrame(data)
-    df.to_csv(stems_path + name, index=False)
-    print()
 
 
 def classify():
@@ -46,10 +31,7 @@ def classify():
         test_data = data[data['Book'] >= 44]['Scripture']
         trains.append(train_data)
         testes[name.replace('.csv', '')] = test_data
-        try:
-            stems_dic[name.replace('.csv', '')] = pd.read_csv(stems_path + name, encoding='utf-8')
-        except pd.errors.EmptyDataError:
-            pass
+        stems_dic[name.replace('.csv', '')] = pd.read_csv(stems_path + name, encoding='utf-8')
 
     prep = PrepData(path)
     lang_dic = prep.get_datasets()
@@ -62,10 +44,29 @@ def classify():
         clf.predict(txt)
 
 
+def stems_analysis():
+    stems = []
+    for label in os.listdir(stems_path):
+        size = len(pd.read_csv(stems_path + label))
+        stems.append((label.replace('.csv', ''), size))
+
+    stems = sorted(stems, key=lambda tup: tup[1])
+
+    labels = []
+    sizes = []
+
+    for tup in stems:
+        labels.append(tup[0])
+        sizes.append(tup[1])
+
+    draw_plot({'x': sizes, 'y': labels})
+
+
 def main():
-    for name in data_list:
-        stem_words(name)
-    classify()
+    """for name in data_list:
+        stem_words(name)"""
+    # classify()
+    stems_analysis()
 
 
 if __name__ == "__main__":
